agent-2-rag/requirements.txt
fastapi
uvicorn[standard]
pydantic
PyJWT
python-dotenv
langgraph
openai
langchain-openai
psycopg2-binary
sqlalchemy